Crate fluence_libp2p

Modules

multihash_serializer

peerid_serializer

provider_serializer

types

Macros

event_polling
Intended to simplify simple polling functions that just return internal events from a internal queue.

generate_swarm_event_type
Generates a type of events produced by Swarm by its name

poll_loop

Structs

PeerId
Identifier of a peer of the network.

RandomPeerId

Functions

build_memory_transport

build_transport
Creates transport that is common for all connections.

remote_multiaddr
Retrieves multiaddr of the remote peer